Move printlog
