+++ src/drivers/helpers/print-log.go
@@ -1,4 +1,4 @@
-package drivers
+package helpers
 
 import (
 	"fmt"
@@ -6,7 +6,8 @@ import (
 	"github.com/fatih/color"
 )
 
-func printLog(message string) {
+// PrintLog prints the given log message in bold.
+func PrintLog(message string) {
 	fmt.Println()
 	_, err := color.New(color.Bold).Println(message)
 	if err != nil {
